pyproject.toml
[tool.poetry]
name = "python-azure-web-app"
version = "0.1.0"
description = ""
authors = ["Sviatoslav Abramov <sviatoslavabramov6@gmail.com>"]
package-mode = false 

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.14"
pydantic = "^2.11.7"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
